chore: rename dir tcp to transport
