+++ speedtest/request.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/showwin/speedtest-go/speedtest/transport"
 	"io"
 	"math"
 	"net/http"
@@ -12,8 +13,6 @@ import (
 	"strings"
 	"sync/atomic"
 	"time"
-
-	"github.com/showwin/speedtest-go/speedtest/tcp"
 )
 
 type (
@@ -280,7 +279,7 @@ func (s *Server) TCPPing(
 		pingDst = s.Host
 	}
 	failTimes := 0
-	client := tcp.NewClient(s.Context.tcpDialer, pingDst)
+	client := transport.NewClient(s.Context.tcpDialer, pingDst)
 	err = client.Connect()
 	if err != nil {
 		return nil, err
